+++ src/neuroconv/tools/nwb_helpers/_dataset_configuration.py
@@ -4,19 +4,12 @@ from typing import Generator, Literal, Union
 import h5py
 import numpy as np
 import zarr
-from hdmf.data_utils import DataChunkIterator, DataIO, GenericDataChunkIterator
-from hdmf.utils import get_data_shape
+from hdmf.data_utils import DataIO
 from hdmf_zarr import NWBZarrIO
-from pynwb import NWBHDF5IO, NWBFile, TimeSeries
+from pynwb import NWBHDF5IO, NWBFile
 from pynwb.base import DynamicTable
 
-from ._models._base_models import DatasetInfo, DatasetIOConfiguration
-from ._models._hdf5_models import HDF5BackendConfiguration, HDF5DatasetIOConfiguration
-from ._models._zarr_models import ZarrBackendConfiguration, ZarrDatasetIOConfiguration
-from ..hdmf import SliceableDataChunkIterator
-
-BACKEND_TO_DATASET_CONFIGURATION = dict(hdf5=HDF5DatasetIOConfiguration, zarr=ZarrDatasetIOConfiguration)
-BACKEND_TO_CONFIGURATION = dict(hdf5=HDF5BackendConfiguration, zarr=ZarrBackendConfiguration)
+from ._configuration_models._base_dataset_io import DatasetIOConfiguration
 
 
 def _get_io_mode(io: Union[NWBHDF5IO, NWBZarrIO]) -> str:
@@ -55,8 +48,9 @@ def get_default_dataset_io_configurations(
     """
     Generate DatasetIOConfiguration objects for wrapping NWB file objects with a specific backend.
 
-    This method automatically detects all objects in an NWB file that can be wrapped in a DataIO. It supports auto-detection
-    of the backend if the NWB file is in append mode, otherwise it requires a backend specification.
+    This method automatically detects all objects in an NWB file that can be wrapped in a DataIO.
+    If the NWB file is in append mode, it supports auto-detection of the backend.
+    Otherwise, it requires a backend specification.
 
     Parameters
     ----------
@@ -70,7 +64,9 @@ def get_default_dataset_io_configurations(
     DatasetIOConfiguration
         A summary of each detected object that can be wrapped in a DataIO.
     """
-    DatasetIOConfigurationClass = BACKEND_TO_DATASET_CONFIGURATION[backend]
+    from ..nwb_helpers import DATASET_IO_CONFIGURATIONS
+
+    DatasetIOConfigurationClass = DATASET_IO_CONFIGURATIONS[backend]
 
     if backend is None and nwbfile.read_io is None:
         raise ValueError(
@@ -146,18 +142,3 @@ def get_default_dataset_io_configurations(
                 )
 
                 yield dataset_io_configuration
-
-
-def get_default_backend_configuration(
-    nwbfile: NWBFile, backend: Literal["hdf5", "zarr"]
-) -> Union[HDF5BackendConfiguration, ZarrBackendConfiguration]:
-    """Fill a default backend configuration to serve as a starting point for further customization."""
-    BackendConfigurationClass = BACKEND_TO_CONFIGURATION[backend]
-    default_dataset_configurations = get_default_dataset_io_configurations(nwbfile=nwbfile, backend=backend)
-    dataset_configurations = {
-        default_dataset_configuration.dataset_info.location: default_dataset_configuration
-        for default_dataset_configuration in default_dataset_configurations
-    }
-
-    backend_configuration = BackendConfigurationClass(dataset_configurations=dataset_configurations)
-    return backend_configuration